requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.6.0
    # via django
django==3.2.19
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.4
    # via django
